+++ src/GutterDecorator.ts
@@ -80,7 +80,7 @@ export class GutterDecorator {
       highlightRanges.push(new vscode.Range(note.from, 0, note.to, Number.MAX_SAFE_INTEGER));
     }
 
-    const showRuler = vscode.workspace.getConfiguration('smartnotes').get<boolean>('showOverviewRuler', true);
+    const showRuler = vscode.workspace.getConfiguration('smartnotes').get<boolean>('showOverviewRuler', false);
     editor.setDecorations(this.gutterType, gutterRanges);
     editor.setDecorations(this.highlightType, showRuler ? highlightRanges : []);
   }